% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCTR.R
\name{aCTR}
\alias{aCTR}
\title{D3 Visualizations: Collapsible Tree}
\source{
D3.js was created by Michael Bostock. See \url{http://d3js.org/}
}
\usage{
aCTR(data, width = 500, height = 300, HT = 20, WD = 6, maxChar = 50,
  maxsize = 20, minimum_distance = 21,
  top_bar = "BRAND, SEGMENT, BASE SIZE, SUB1, SUB2, SUB3",
  color1 = "lightsteelblue", color2 = "#fff", color3 = "steelblue",
  color4 = "#ccc")
}
\arguments{
\item{data}{the json file being used for the visualizations.}

\item{width}{width for the graph's frame area (in pixels) - default is null.}

\item{height}{height for the graph's frame area (in pixels) - default is null.}

\item{HT}{what is the maximum number of leaves in a column}

\item{WD}{what is the maximum depth of the tree}

\item{maxChar}{what is the maximum number of characters in a nodes text}

\item{maxsize}{what is the root nodes size}

\item{color1}{the color of the circles which still have children - default lightbluesteel}

\item{color2}{the color of the circles whose children are already shown, or that do not have children - default black}

\item{color3}{the color of the outside of the circles - default steelblue}

\item{color4}{the color of the lines connecting the circles - default grey}
}
\description{
Creates a collapsible tree given an appropraite json file.
}
\examples{
\dontrun{
# load in an appropriate json file.
# Such as \\url{https://gist.github.com/mbostock/1093025#file-flare-json}
# we will call this json.json
aCTR(json.json) # This should reproduce the Mike Bostock's Example
aCTR(json.json, color1 = "blue", color2 = "red", color3 = "green", color4 = "black")
# Here we change around the colors of the visualization.
}

}
