% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIT.R
\name{CIT}
\alias{CIT}
\title{D3 Visualization: Collapsible Indented Tree}
\source{
D3.js was created by Michael Bostock. See \url{http://d3js.org/}
}
\usage{
CIT(data, height = NULL, width = NULL, color1 = "#bd3182",
  color2 = "#31bd6c", color3 = "#c6dbef")
}
\arguments{
\item{data}{the json file being used for the visualizations.}

\item{height}{height for the graph's frame area (in pixels) - default is null.}

\item{width}{width for the graph's frame area (in pixels) - default is null.}

\item{color1}{the color of the bars when the branch is collapsed}

\item{color2}{the color of the bars when the branch is not collapsed}

\item{color3}{the color of the bars that represent the children}
}
\description{
Creates a collapsible indented tree.
}
\examples{
\dontrun{
# load in an appropriate json file. 
# Such as \\url{https://gist.github.com/mbostock/1093025#file-flare-json}
# we will call this json.json
CIT(json.json) # This should reproduce the Mike Bostock's Example
CIT(json.json, color1 = "blue", color2 = "red", color3 = "green") 
# Here we change around the colors of the visualization.
}

}
