% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTR.R
\name{CTR}
\alias{CTR}
\title{D3 Visualizations: Collapsible Tree}
\source{
D3.js was created by Michael Bostock. See \url{http://d3js.org/}
}
\usage{
CTR(data, width = NULL, height = NULL, color1 = "lightsteelblue",
  color2 = "#fff", color3 = "steelblue", color4 = "#ccc")
}
\arguments{
\item{data}{the json file being used for the visualizations.}

\item{width}{width for the graph's frame area (in pixels) - default is null.}

\item{height}{height for the graph's frame area (in pixels) - default is null.}

\item{color1}{the color of the circles which still have children - default lightbluesteel}

\item{color2}{the color of the circles whose children are already shown, or that do not have children - default black}

\item{color3}{the color of the outside of the circles - default steelblue}

\item{color4}{the color of the lines connecting the circles - default grey}
}
\description{
Creates a collapsible tree given an appropraite json file.
}
\examples{
\dontrun{
# load in an appropriate json file. 
# Such as \\url{https://gist.github.com/mbostock/1093025#file-flare-json}
# we will call this json.json
CTR(json.json) # This should reproduce the Mike Bostock's Example
CTR(json.json, color1 = "blue", color2 = "red", color3 = "green", color4 = "black") 
# Here we change around the colors of the visualization.
}

}
