% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ICICLE.R
\name{ICICLE}
\alias{ICICLE}
\title{D3 Icicile Plot}
\source{
D3.js was created by Michael Bostock. See \url{http://d3js.org/}
}
\usage{
ICICLE(data, width = 960, height = 500, color = "d3.scale.category20()",
  mousemove = NULL)
}
\arguments{
\item{data}{A character string, one line, containing the JSON data to be parsed.}

\item{width}{Width for the graph's frame area (in pixels) - default is null.}

\item{height}{Height for the graph's frame area (in pixels) - default is null.}

\item{color}{Place to add custom color for the ICICLE - default is d3.scale.category20().}

\item{mousemove}{Variable to add custom JS for a tooltip that appears on hover. The example uses this with DIV tags.}
}
\description{
Create an Icicle D3 Visualization
}
\examples{
\dontrun{
ICICLE("a")
}
}

